+++ pytket/tests/predicates_test.py
@@ -1050,7 +1050,6 @@ def test_greedy_pauli_synth() -> None:
     range_predicate = RangePredicateOp(3, 0, 6)
     set_bits = SetBitsOp([True, True])
     multi_bit = MultiBitOp(set_bits, 2)
-    exp = Bit(2) & Bit(3)
     eq_pred_values = [True, False, False, True]
     and_values = [bool(i) for i in [0, 0, 0, 1]]
     pg1 = PauliExpBox([Pauli.X, Pauli.Z], 0.3)
@@ -1058,7 +1057,6 @@ def test_greedy_pauli_synth() -> None:
     circ.add_pauliexpbox(pg1, [0, 1])
     circ.add_gate(multi_bit, [0, 1, 2, 3])
     circ.add_gate(range_predicate, [0, 1, 2, 3])
-    circ.add_classicalexpbox_bit(exp, [Bit(0)])
     circ.add_c_predicate(eq_pred_values, [0, 1], 2, "EQ")
     circ.add_c_modifier(and_values, [1], 2)
     circ._add_wasm("funcname", "wasmfileuid", [1, 1], [], [Bit(0), Bit(1)], [0])
@@ -1073,7 +1071,6 @@ def test_greedy_pauli_synth() -> None:
     d.H(0)
     d.add_gate(multi_bit, [0, 1, 2, 3])
     d.add_gate(range_predicate, [0, 1, 2, 3])
-    d.add_classicalexpbox_bit(exp, [Bit(0)])
     d.add_c_predicate(eq_pred_values, [0, 1], 2, "EQ")
     d.add_c_modifier(and_values, [1], 2)
     d._add_wasm("funcname", "wasmfileuid", [1, 1], [], [Bit(0), Bit(1)], [0])