+++ pytket/tests/qasm_test.py
@@ -454,7 +454,7 @@ def test_h1_rzz() -> None:
 def test_extended_qasm() -> None:
     fname = str(curr_file_path / "qasm_test_files/test17.qasm")
     out_fname = str(curr_file_path / "qasm_test_files/test17_output.qasm")
-    c = circuit_from_qasm_wasm(fname, "testfile.wasm", use_clexpr=True)
+    c = circuit_from_qasm_wasm(fname, "testfile.wasm")
 
     out_qasm = circuit_to_qasm_str(c, "hqslib1")
     with open(out_fname) as f:
@@ -468,14 +468,11 @@ def test_extended_qasm() -> None:
         DecomposeClassicalExp().apply(c)
 
 
-@pytest.mark.parametrize("use_clexpr", [True, False])
-def test_decomposable_extended(use_clexpr: bool) -> None:
+def test_decomposable_extended() -> None:
     fname = str(curr_file_path / "qasm_test_files/test18.qasm")
     out_fname = str(curr_file_path / "qasm_test_files/test18_output.qasm")
 
-    c = circuit_from_qasm_wasm(
-        fname, "testfile.wasm", maxwidth=64, use_clexpr=use_clexpr
-    )
+    c = circuit_from_qasm_wasm(fname, "testfile.wasm", maxwidth=64)
     DecomposeClassicalExp().apply(c)
 
     out_qasm = circuit_to_qasm_str(c, "hqslib1", maxwidth=64)
@@ -690,11 +687,10 @@ def test_qasm_phase() -> None:
         assert c1 == c0
 
 
-@pytest.mark.parametrize("use_clexpr", [True, False])
-def test_CopyBits(use_clexpr: bool) -> None:
+def test_CopyBits() -> None:
     input_qasm = """OPENQASM 2.0;\ninclude "hqslib1.inc";\n\ncreg c0[1];
 creg c1[3];\nc0[0] = c1[1];\n"""
-    c = circuit_from_qasm_str(input_qasm, use_clexpr=use_clexpr)
+    c = circuit_from_qasm_str(input_qasm)
     result_circ_qasm = circuit_to_qasm_str(c, "hqslib1")
     assert input_qasm == result_circ_qasm
 
@@ -868,8 +864,7 @@ measure q[0] -> c[32];"""
     assert len(circ_out.bits) == 33
 
 
-@pytest.mark.parametrize("use_clexpr", [True, False])
-def test_classical_expbox_arg_order(use_clexpr: bool) -> None:
+def test_classical_expbox_arg_order() -> None:
     qasm = """
     OPENQASM 2.0;
     include "hqslib1.inc";
@@ -884,7 +879,7 @@ def test_classical_expbox_arg_order(use_clexpr: bool) -> None:
     c = a ^ b | d;
     """
 
-    circ = circuit_from_qasm_str(qasm, use_clexpr=use_clexpr)
+    circ = circuit_from_qasm_str(qasm)
     args = circ.get_commands()[0].args
     expected_symbol_order = ["a", "b", "d", "c"]
     expected_index_order = [0, 1, 2, 3]
@@ -1286,13 +1281,13 @@ if __name__ == "__main__":
     test_hqs_conditional_params()
     test_barrier()
     test_barrier_2()
-    test_decomposable_extended(True)
-    test_decomposable_extended(False)
+    test_decomposable_extended()
+    test_decomposable_extended()
     test_alternate_encoding()
     test_header_stops_gate_definition()
     test_tk2_definition()
     test_rxxyyzz_conversion()
-    test_classical_expbox_arg_order(True)
-    test_classical_expbox_arg_order(False)
+    test_classical_expbox_arg_order()
+    test_classical_expbox_arg_order()
     test_register_name_check()
     test_existing_name_conversion()