+++ tket/src/Predicates/CompilerPass.cpp
@@ -428,8 +428,6 @@ PassPtr deserialise(
       pp = SynthesiseTK();
     } else if (passname == "SynthesiseTket") {
       pp = SynthesiseTket();
-    } else if (passname == "SynthesiseUMD") {
-      pp = SynthesiseUMD();
     } else if (passname == "SquashTK1") {
       pp = SquashTK1();
     } else if (passname == "SquashRzPhasedX") {
@@ -471,8 +469,7 @@ PassPtr deserialise(
       pp = gen_euler_pass(q, p, s);
     } else if (passname == "FlattenRelabelRegistersPass") {
       pp = gen_flatten_relabel_registers_pass(
-          content.at("label").get<std::string>(),
-          content.at("relabel_classical_expressions").get<bool>());
+          content.at("label").get<std::string>());
     } else if (passname == "RoutingPass") {
       Architecture arc = content.at("architecture").get<Architecture>();
       std::vector<RoutingMethodPtr> con = content.at("routing_config");